# docs/swagger.yaml
basePath: /api/v1
definitions:
  api.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  profile.Profile:
    properties:
      lastAccess:
        type: string
      name:
        type: string
      picture:
        type: string
      pictureLocked:
        type: boolean
    type: object
  profile.UpdateProfileRequest:
    properties:
      forceUpdate:
        type: boolean
      name:
        type: string
      picture:
        type: string
    required:
    - name
    type: object
  settings.Settings:
    properties:
      autoBuyCards:
        type: boolean
    type: object
info:
  contact: {}
  description: An API for playing the card game called 110. 110 is a game based on
    the game 25 and is played primarily in Ireland
  title: Cards 110 API
  version: 8.0.0
paths:
  /profile:
    get:
      description: Returns the user's profile.
      operationId: get-profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/profile.Profile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - Bearer: []
      tags:
      - Profile
    put:
      consumes:
      - application/json
      description: Updates the user's profile or creates it if it doesn't exist.
      operationId: update-profile
      parameters:
      - description: Profile
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/profile.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/profile.Profile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - Bearer: []
      tags:
      - Profile
  /profile/has:
    get:
      description: Returns a boolean indicating if the user has a profile or not.
      operationId: has-profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - Bearer: []
      tags:
      - Profile
  /settings:
    get:
      description: Returns the user's settings.
      operationId: get-settings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/settings.Settings'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - Bearer: []
      tags:
      - Settings
    put:
      consumes:
      - application/json
      description: Saves the user's settings.
      operationId: save-settings
      parameters:
      - description: Settings
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/settings.Settings'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/settings.Settings'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - Bearer: []
      tags:
      - Settings
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"